+++ airbyte-cdk/python/airbyte_cdk/sources/utils/slice_logger.py
