+++ airbyte-cdk/python/unit_tests/sources/utils/test_slice_logger.py
@@ -7,7 +7,7 @@ import logging
 import pytest
 from airbyte_cdk.models import AirbyteLogMessage, AirbyteMessage, Level
 from airbyte_cdk.models import Type as MessageType
-from airbyte_cdk.sources.utils.slice_logging import DebugSliceLogger
+from airbyte_cdk.sources.utils.slice_logger import DebugSliceLogger
 
 
 @pytest.mark.parametrize(
